fix: AP News
