+++ lib/router.js
@@ -2505,7 +2505,6 @@ router.get('/gbcc/trust', require('./routes/gbcc/trust'));
 
 // Associated Press
 router.get('/apnews/topics/:topic', require('./routes/apnews/topics'));
-router.get('/apnews/topics2/:topic', require('./routes/apnews/topics2'));
 
 // CBC
 router.get('/cbc/topics/:topic?', require('./routes/cbc/topics'));
